requirements.txt
Flask==3.0.3
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
python-whois==0.8.0
googlesearch-python==1.2.5
lxml==5.2.2
